refactor(cli): Move subommands to their own files
