+++ cli/market/common.py
@@ -1,0 +1,27 @@
+from __future__ import annotations
+
+from pathlib import Path
+import os
+import subprocess
+
+import typer
+
+REPO_ROOT = Path(__file__).resolve().parents[2]
+
+
+def run_step(
+    label: str,
+    cmd: list[str],
+    cwd: Path,
+    extra_env: dict[str, str] | None = None,
+) -> None:
+    typer.echo(f"==> {label} at {cwd}")
+    env = os.environ.copy()
+    venv_path = cwd / ".venv"
+    venv_bin = venv_path / "bin"
+    if venv_bin.exists():
+        env["VIRTUAL_ENV"] = str(venv_path)
+        env["PATH"] = f"{venv_bin}{os.pathsep}{env.get('PATH', '')}"
+    if extra_env:
+        env.update(extra_env)
+    subprocess.run(cmd, cwd=cwd, check=True, env=env)
