+++ cli/market/groups/__init__.py
@@ -1,0 +1,2 @@
+"""Typer command groups for the market CLI."""
+
